+++ src/extension.test.ts
@@ -109,6 +109,7 @@ describe("extension", () => {
                 defaultUrl: "url",
                 hostname: "localhost",
                 port: 9222,
+                timeout: 10000,
                 useHttps: false,
                 userDataDir: "profile",
             });
@@ -178,6 +179,7 @@ describe("extension", () => {
                     getRemoteEndpointSettings: jest.fn().mockReturnValue({
                         hostname: "hostname",
                         port: "port",
+                        timeout: 10000,
                         useHttps: false,
                     }),
                     getRuntimeConfig: jest.fn().mockReturnValue(fakeRuntimeConfig),
@@ -258,6 +260,7 @@ describe("extension", () => {
                 webSocketDebuggerUrl: expectedWS,
             } as IRemoteTargetJson;
 
+            mocks.utils.getListOfTargets!.mockResolvedValue([target]),
             mocks.utils.fixRemoteWebSocket!.mockReturnValueOnce(target);
 
             const expectedContext = createFakeExtensionContext();
@@ -280,6 +283,7 @@ describe("extension", () => {
                 webSocketDebuggerUrl: expectedWS,
             } as IRemoteTargetJson;
 
+            mocks.utils.getListOfTargets!.mockResolvedValue([target]),
             mocks.utils.fixRemoteWebSocket!.mockReturnValueOnce(target);
 
             const expectedContext = createFakeExtensionContext();
@@ -295,6 +299,29 @@ describe("extension", () => {
             // Reverse the mismatched slashes
             target.url = expectedUrl;
             await newExtension.attach(expectedContext, `${expectedUrl}/`);
+            expect(mocks.panel.DevToolsPanel.createOrShow).toHaveBeenCalledWith(
+                expectedContext,
+                mockTelemetry,
+                expectedWS,
+                fakeRuntimeConfig,
+            );
+        });
+
+        it("opens devtools against given filter", async () => {
+            const expectedUrl = "http://target:9222";
+            const expectedWS = "ws://target:9222";
+            target = {
+                title: "title",
+                url: expectedUrl,
+                webSocketDebuggerUrl: expectedWS,
+            } as IRemoteTargetJson;
+
+            mocks.utils.getListOfTargets!.mockResolvedValue([target]),
+            mocks.utils.fixRemoteWebSocket!.mockReturnValueOnce(target);
+
+            const expectedContext = createFakeExtensionContext();
+            const newExtension = await import("./extension");
+            await newExtension.attach(expectedContext, "http://*");
             expect(mocks.panel.DevToolsPanel.createOrShow).toHaveBeenCalledWith(
                 expectedContext,
                 mockTelemetry,
@@ -344,6 +371,7 @@ describe("extension", () => {
                 getRemoteEndpointSettings: jest.fn().mockReturnValue({
                     hostname: "hostname",
                     port: "port",
+                    timeout: 10000,
                     useHttps: false,
                 }),
                 getRuntimeConfig: jest.fn().mockReturnValue(fakeRuntimeConfig),