+++ CHANGELOG.md
@@ -1,3 +1,19 @@
+## 1.0.4
+* Minor change to icon
+
+## 1.0.3
+* Added Debugger for Microsoft Edge integration - [PR #96](https://github.com/microsoft/vscode-edge-devtools/pull/96)
+* Fixed content-security-policy - [PR #95](https://github.com/microsoft/vscode-edge-devtools/pull/95)
+* Fixed closing browser now also closes elements tool - [PR #89](https://github.com/microsoft/vscode-edge-devtools/pull/89)
+* Bumped elements tool version to `78.0.273.0`
+* Updated readme
+
+## 1.0.2
+* Added open links in editor feature - [PR #83](https://github.com/microsoft/vscode-edge-devtools/pull/83)
+* Added new settings and launch.json configs for sourcemaps - See [Sourcemaps](https://github.com/microsoft/vscode-edge-devtools#sourcemaps)
+* Bumped elements tool version to `77.0.223.0`
+* Updated readme
+
 ## 1.0.1
 * Fixed launch procedure so that remote debugging will be enabled correctly - [PR #67](https://github.com/microsoft/vscode-edge-devtools/pull/67)
 * Updated readme